requirements.txt
Flask==0.10.1
Flask-Cors==1.9.0
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==0.3.0
Werkzeug==0.9.6
argparse==1.2.1
gunicorn==19.1.1
itsdangerous==0.24
six==1.8.0
wsgiref==0.1.2
